requirements.txt
accelerate
peft~=0.5.0
bitsandbytes
transformers~=4.33.1
trl~=0.7.1
datasets~=2.14.5
langchain~=0.0.285
xformers
python-dotenv~=1.0.0
torch~=2.0.1
pandas~=2.1.0
google~=3.0.0
google-api-python-client
gspread~=5.11.1
peft
scipy
PyYAML~=6.0.1
google-api-core